Rename `fieldtype` package to `field.type`
