+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/field/type/FieldTypes.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.compiler.fieldtype;
+package io.spine.tools.compiler.field.type;
 
 import com.google.protobuf.DescriptorProtos.FieldDescriptorProto;
