+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/field/type/SingularFieldType.java
@@ -17,38 +17,32 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-package io.spine.tools.compiler.fieldtype;
+package io.spine.tools.compiler.field.type;
 
 import com.squareup.javapoet.ClassName;
-import com.squareup.javapoet.ParameterizedTypeName;
 import com.squareup.javapoet.TypeName;
+import io.spine.code.java.PrimitiveType;
 
-import java.util.Map;
+import java.util.Optional;
 
 /**
- * Represents map {@linkplain FieldType field type}.
+ * Represents singular {@linkplain FieldType field type}.
+ *
+ * @author Dmytro Grankin
  */
-public class MapFieldType implements FieldType {
+public class SingularFieldType implements FieldType {
 
-    private static final String SETTER_PREFIX = "putAll";
+    private static final String SETTER_PREFIX = "set";
 
     private final TypeName typeName;
-    private final TypeName keyTypeName;
-    private final TypeName valueTypeName;
 
     /**
-     * Constructs the {@link MapFieldType} based on
-     * the key and the value type names.
+     * Constructs the {@link SingularFieldType} based on field type name.
      *
-     * @param entryTypeNames the entry containing the key and the value type names.
+     * @param name the field type name
      */
-    MapFieldType(Map.Entry<TypeName, TypeName> entryTypeNames) {
-        this.keyTypeName = boxIfPrimitive(entryTypeNames.getKey());
-        this.valueTypeName = boxIfPrimitive(entryTypeNames.getValue());
-        this.typeName = ParameterizedTypeName.get(ClassName.get(Map.class),
-                                                  keyTypeName,
-                                                  valueTypeName);
+    SingularFieldType(String name) {
+        this.typeName = constructTypeNameFor(name);
     }
 
     /**
@@ -59,19 +53,11 @@ public class MapFieldType implements FieldType {
         return typeName;
     }
 
-    public TypeName getKeyTypeName() {
-        return keyTypeName;
-    }
-
-    public TypeName getValueTypeName() {
-        return valueTypeName;
-    }
-
     /**
-     * Returns "putAll" setter prefix,
-     * used to initialize a map field using a protobuf message builder.
+     * Returns "set" setter prefix,
+     * used to initialize a singular field using a protobuf message builder.
      *
-     * <p>Call should be like `builder.putAllFieldName({@link Map})`.
+     * Call should be like `builder.setFieldName(FieldType)`.
      *
      * @return {@inheritDoc}
      */
@@ -80,12 +66,14 @@ public class MapFieldType implements FieldType {
         return SETTER_PREFIX;
     }
 
-    private static TypeName boxIfPrimitive(TypeName typeName) {
-        if (typeName.isPrimitive()) {
-            return typeName.box();
-        }
+    private static TypeName constructTypeNameFor(String name) {
+        Optional<? extends Class<?>> boxedScalarPrimitive =
+                PrimitiveType.getWrapperClass(name);
 
-        return typeName;
+        return boxedScalarPrimitive.isPresent()
+               ? TypeName.get(boxedScalarPrimitive.get())
+                         .unbox()
+               : ClassName.bestGuess(name);
     }
 
     @Override
