+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/validation/AbstractMethodConstructorBuilder.java
@@ -23,7 +23,7 @@ package io.spine.tools.compiler.validation;
 import com.google.protobuf.DescriptorProtos.FieldDescriptorProto;
 import com.squareup.javapoet.ClassName;
 import io.spine.tools.compiler.MessageTypeCache;
-import io.spine.tools.compiler.fieldtype.FieldType;
+import io.spine.tools.compiler.field.type.FieldType;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import static com.google.common.base.Preconditions.checkArgument;
