+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/validation/SingularFieldMethodConstructor.java
@@ -29,10 +29,9 @@ import io.spine.base.ConversionException;
 import io.spine.code.proto.FieldName;
 import io.spine.logging.Logging;
 import io.spine.tools.compiler.MessageTypeCache;
-import io.spine.tools.compiler.fieldtype.FieldType;
+import io.spine.tools.compiler.field.type.FieldType;
 import io.spine.validate.ValidationException;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.lang.model.element.Modifier;
 import java.util.Collection;
