+++ tools/model-compiler/src/test/java/io/spine/tools/compiler/field/type/FieldTypesShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.compiler.fieldtype;
+package io.spine.tools.compiler.field.type;
 
 import com.google.common.testing.NullPointerTester;
 import org.junit.Test;
